Fix upstream dockerfile and add 'by hand' ctrfile
The podmanimage/upstream/Dockerfile had two rpms in its
build procedure that are no longer available.  The atomic-registries
has been removed and the md2man has been renamed.  In addtion
conmon was not being installed and I've added that.

I've been using a Containerfile to build or rebuild a
specific version of the podmanimage stored in the stable
repository with a version tag.  As the other Containerfiles
have been updated by others, and in case anyone else needs
to build it, I've added it to the repo and have also updated
the readme.md.

FWIW, the builds in the quay.io/podman/upstream have been failing for a while due to missing rpms.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>